refactor(planning_test_utils): remove route_handler dependencies
