+++ planning/autoware_planning_test_manager/src/autoware_planning_test_manager.cpp
@@ -15,6 +15,9 @@
 #include "motion_utils/trajectory/conversion.hpp"
 
 #include <autoware_planning_test_manager/autoware_planning_test_manager.hpp>
+#include <autoware_planning_test_manager/autoware_planning_test_manager_utils.hpp>
+#include <tier4_autoware_utils/geometry/geometry.hpp>
+
 #include <planning_test_utils/planning_test_utils.hpp>
 
 namespace planning_test_utils
@@ -115,7 +118,7 @@ void PlanningInterfaceTestManager::publishInitialPose(
   if (module_name == ModuleName::START_PLANNER) {
     test_utils::publishToTargetNode(
       test_node_, target_node, topic_name, initial_pose_pub_,
-      test_utils::makeInitialPoseFromLaneId(10291));
+      autoware_planning_test_manager::utils::makeInitialPoseFromLaneId(10291));
   } else {
     test_utils::publishToTargetNode(
       test_node_, target_node, topic_name, initial_pose_pub_, test_utils::makeInitialPose(shift));
@@ -256,7 +259,7 @@ void PlanningInterfaceTestManager::publishBehaviorNominalRoute(
   if (module_name == ModuleName::START_PLANNER) {
     test_utils::publishToTargetNode(
       test_node_, target_node, topic_name, behavior_normal_route_pub_,
-      test_utils::makeBehaviorRouteFromLaneId(10291, 10333), 5);
+      autoware_planning_test_manager::utils::makeBehaviorRouteFromLaneId(10291, 10333), 5);
   } else {
     test_utils::publishToTargetNode(
       test_node_, target_node, topic_name, behavior_normal_route_pub_,
