+++ planning/planning_test_utils/package.xml
@@ -25,10 +25,8 @@
   <depend>component_interface_utils</depend>
   <depend>lanelet2_extension</depend>
   <depend>lanelet2_io</depend>
-  <depend>motion_utils</depend>
   <depend>nav_msgs</depend>
   <depend>rclcpp</depend>
-  <depend>route_handler</depend>
   <depend>tf2_msgs</depend>
   <depend>tf2_ros</depend>
   <depend>tier4_api_msgs</depend>
